+++ hybrid-compute/offchain/check_kyc/check_kyc_offchain.py
@@ -3,6 +3,11 @@ from web3 import Web3
 from eth_abi import abi as ethabi
 from hybrid_compute_sdk.server import HybridComputeSDK
 
+def get_handlers():
+    """Return the method signatures and the associated handlers"""
+    print("--> checkkyc(string)")
+    return [("checkkyc(string)",        offchain_checkkyc)]
+
 validWallets = ["0x123"]
 
 def offchain_checkkyc(ver, sk, src_addr, src_nonce, oo_nonce, payload, *args):
