+++ hybrid-compute/offchain/handlers/auction_system_offchain.py
@@ -1,0 +1,1 @@
+../auction_system/auction_system_offchain.py
