+++ hybrid-compute/offchain/handlers/check_kyc_offchain.py
@@ -1,0 +1,1 @@
+../check_kyc/check_kyc_offchain.py
