+++ hybrid-compute/offchain/handlers/ramble_offchain.py
@@ -1,0 +1,1 @@
+../ramble/ramble_offchain.py
