+++ hybrid-compute/offchain/handlers/sports_betting_offchain.py
@@ -1,0 +1,1 @@
+../sports_betting/sports_betting_offchain.py
