+++ hybrid-compute/offchain/handlers/vrf_offchain.py
@@ -1,0 +1,1 @@
+../vrf/vrf_offchain.py
